service-marketplace/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25

Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
email-validator==2.2.0

gunicorn==23.0.0
python-dotenv==1.0.1